backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
greenlet==3.1.1
pymysql
jinja2
python-dotenv
pydantic
